+++ examples/nextjs-relay/src/lib/wundergraph/index.ts
@@ -1,0 +1,8 @@
+import { createWunderGraphRelayApp } from '../../../../../packages/react-relay';
+import { createClient } from '../../../.wundergraph/generated/nextjs/client';
+
+const client = createClient();
+
+export const { WunderGraphRelayProvider, useLiveQuery, fetchWunderGraphSSRQuery } = createWunderGraphRelayApp({
+	client,
+});